Renamed constants file to helpers
