+++ lib/components/Editor/CustomExtensions/BubbleMenu/ImageOptions.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { getImageMenuOptions } from "./constants";
+import { getImageMenuOptions } from "./helpers";
 import Option from "./Option";
 
 const ImageOptions = ({
