+++ lib/components/Editor/CustomExtensions/BubbleMenu/TextOptions.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { getTextMenuOptions } from "./constants";
+import { getTextMenuOptions } from "./helpers";
 import LinkOption from "./LinkOption";
 import Option from "./Option";
 
@@ -13,9 +13,8 @@ const TextOptions = ({
 }) => {
   const handleAnimateInvalidLink = () => {
     setIsInvalidLink(true);
-    const timer = setTimeout(() => {
+    setTimeout(() => {
       setIsInvalidLink(false);
-      clearTimeout(timer);
     }, 1000);
   };
